server/requirements.txt
flask
PyPDF2
openai
python-dotenv
reportlab
pdfminer
pdfminer.six
flask_cors
mysql-connector-python
bcrypt
flask_pymongo